deployment/start/requirements.txt
azure-functions
requests
onnxruntime
onnx
torch
PyYAML
transformers
